Finished get followers and following
Both get functions return a set. Then a user to the unfollowing list if
they exist in the following set but not in the followers set. The
program then prints their '@' to let the user know.

Can also add functionality such as using a dictionary instead of a set,
where the key will be the '@' and the value can be another dictionary
which can contain information about the following/follower.