Changed logic for getting unfollowers
Previously, the `GetUnfollowers()` function would compute/get a list of
all users that are in the "following" list, and then get another list of
all users that are in the "followers" list. Then check whether each user
in the "following" set is not in the "followers" set, if so, then that
user is considered an unfollower. Now, just traverse over the
"following" list and if a user is following back, then their user cell
would contain a "Follows you" text to the right of their '@'. It's
faster now compared to the previous approach. Although, may need to
revert back if their values change/unable to get the value to the right
of the '@' of each user.